INITIAL COMMIT OF THE DATAPARSE, MODELS
This is an initial setup and commit of the dataparse and models
modules within the repo.

In dataparse, we initialize a reader, which is meant to read in all
the data necessary, whether by scraping through text files, or reading
through excel files.

In models, we define the different sql schemas we need to store the
varying levels of the data.