PROTOTYPED MODELS WITH NOTEBOOK WITH PARSER
mvp of a parser for reading patdata table from an excel file.

to Do:
- finalize assumption that patdata has all necessary data.
Possibly add the 'clinical notes', 'date of first recurrence',
'post op progress info'
- add parser for chan xyz data, reg xyz data, connectivity data
Is it necessary? Perhaps we can just store the connectivity.zip files.

notebook related to connecting the parser with model creation.

to do:
- add string parser to summarize the information from excel files.

Added file structure for logging, and connecting to db.
Created a directory for the db to be created '__db__'